+++ protected/humhub/modules/ui/mail/DefaultMailStyle.php
@@ -9,5 +9,10 @@ namespace humhub\modules\ui\mail;
  */
 class DefaultMailStyle
 {
-    public const DEFAULT_FONT_FAMILY = "'Open Sans',Arial,Tahoma,Helvetica,sans-serif";
+    public const DEFAULT_FONT_FAMILY = "-apple-system, BlinkMacSystemFont, 'Segoe UI', ’Helvetica Neue’, Helvetica, Arial, sans-serif";
+    public const DEFAULT_BACKGROUND_COLOR_MAIN = '#fff';
+    public const DEFAULT_BACKGROUND_COLOR_PAGE = '#ededed';
+    public const DEFAULT_COLOR_PRIMARY = '#708fa0';
+    public const DEFAULT_COLOR_INFO = '#6fdbe8';
+    public const DEFAULT_TEXT_COLOR_SOFT2 = '#aeaeae';
 }