+++ protected/humhub/widgets/mails/views/mailButton.php
@@ -7,6 +7,7 @@
  */
 
 use humhub\components\View;
+use humhub\helpers\ScssHelper;
 use humhub\modules\ui\mail\DefaultMailStyle;
 
 /* @var $this View */
@@ -17,7 +18,7 @@ use humhub\modules\ui\mail\DefaultMailStyle;
 <td width="auto"  align="center" valign="middle" height="32"
     style="background-color:<?= $color ?>; border-radius:5px; background-clip: padding-box;font-size:14px; font-family:<?= $this->theme->variable('mail-font-family', DefaultMailStyle::DEFAULT_FONT_FAMILY) ?>; text-align:center;font-weight: 600; padding: 5px 30px">
     <span>
-        <a href="<?= $url ?>" style="text-decoration: none; color: <?= $this->theme->variable('button-text-color', '#fff') ?>; font-weight: 300">
+        <a href="<?= $url ?>" style="text-decoration: none; color: <?= ScssHelper::getColorContrast($color) ?>; font-weight: 400">
             <?= $text ?>
         </a>
     </span>